Fix incorrect initial positions for replicated entities
When an entity is replicated, physics initialization typically happens before transform has been replicated, so the world position at the time of physics initialization is not useful for resetting the position.

In any case, if it's owned remotely, the remote should be managing the position and this client should just shadow that position.